+++ src/test/java/org/morphix/reflection/FieldsGetTest.java
@@ -28,7 +28,7 @@ import org.morphix.reflection.testdata.B;
  *
  * @author Radu Sebastian LAZIN
  */
-class FieldsGetFieldValueTest {
+class FieldsGetTest {
 
 	@Test
 	void shouldReturnFieldValue() throws Exception {
@@ -70,5 +70,28 @@ class FieldsGetFieldValueTest {
 
 		assertThat(e.getMessage(), equalTo("Could not get field " + A.FieldName.B));
 		assertThat(e.getCause(), instanceOf(IllegalArgumentException.class));
+	}
+
+	@Test
+	void shouldRetrieveFieldValueByGetter() {
+		Integer result = Fields.get(new D(), "x");
+
+		assertThat(result, equalTo(2));
+	}
+
+	@Test
+	void shouldRetrieveFieldValueByField() {
+		Integer result = Fields.get(new D(), "y");
+
+		assertThat(result, equalTo(3));
+	}
+
+	public static class D {
+		Integer x = 2;
+		Integer y = 3;
+
+		public Integer getX() {
+			return x;
+		}
 	}
 }